src/com/android/car/settings/common/NoSetupPreferenceController.java
/*
 * Copyright (C) 2018 The Android Open Source Project
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.android.car.settings.common;

import android.car.drivingstate.CarUxRestrictions;
import android.content.Context;

/**
 * Controller that is available when {@link CarUxRestrictions#UX_RESTRICTIONS_NO_SETUP} is not
 * active.
 */
public class NoSetupPreferenceController extends BasePreferenceController {

    public NoSetupPreferenceController(Context context, String preferenceKey) {
        super(context, preferenceKey);
    }

    @Override
    public int getAvailabilityStatus() {
        return AVAILABLE;
    }

    @Override
    protected boolean canBeShownWithRestrictions(CarUxRestrictions restrictionInfo) {
        return !CarUxRestrictionsHelper.isNoSetup(restrictionInfo);
    }
}
